% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_6.R
\name{histogram}
\alias{histogram}
\title{Histogram of residuals}
\usage{
histogram(res)
}
\arguments{
\item{res}{A \code{vector} of length \code{n} that represents the residuals.}
}
\description{
Creates a histogram for the residuals.
}
\examples{
histogram(model$res)
}
\author{
Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett
}
