% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_6.R
\name{our_lm}
\alias{our_lm}
\title{Linear Regression}
\usage{
our_lm(y, X, alpha = 0.05)
}
\arguments{
\item{y}{A \code{vector} of length \code{n} that represents the response variable of interest that we would like to explain/predict.}

\item{X}{A \code{matrix} of dimension \code{n x p} that represents the independent (explanatory) variables.}

\item{alpha}{A \code{numeric} (double) used to denote the significance level.}
}
\value{
A \code{list} containing the following attributes:
\describe{
     \item{beta}{Estimated coefficient vector}
     \item{sigma2}{Estimated variance}
     \item{variance_beta}{Estimated variance for each beta}
     \item{r_squared}{Coefficient of determination}
     \item{mallow}{Mallow's \code{C_p}}
     \item{f_stat}{F-statistic}
     \item{p_value}{Determined by F-statistic}
     \item{ci}{Interval of values based on \code{alpha}}
     \item{res}{Model Residuals}
     \item{preds}{Model Predictions}
}
}
\description{
Estimates a vector for the model coefficients. It also computes model evaluation metrics.
}
\examples{
x <- matrix(rnorm(400), ncol = 4) # so this is a 100x4 predictor matrix
y <- rnorm(100)
model <- our_lm(y, x, alpha = 0.05)
model
}
\author{
Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett
}
