% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression_6.R
\name{qq_plot}
\alias{qq_plot}
\title{qq-plot of residuals}
\usage{
qq_plot(res)
}
\arguments{
\item{res}{A \code{vector} of length \code{n} that represents the residuals.}
}
\description{
Creates a qq-plot for the residuals.
}
\examples{
qq_plot(model$res)
}
\author{
Clayton Haley

Ukamaka Victoria Nnyaba

Robert Fett
}
